Reworked syscalls file to support multicore and single core lnic programs, all with arguments
