+++ tests/lnic-send-recv.c
@@ -4,7 +4,7 @@
 
 #include "lnic.h"
 
-#define NUM_MSG_WORDS 704
+#define NUM_MSG_WORDS 10
 
 uint32_t get_dst_ip(uint32_t nic_ip_addr) {
     if (nic_ip_addr == 0x0a000021) {
